Add permisions for groups Add language

+++ index.php
@@ -17,6 +17,7 @@
 					<tbody>
  <?php
 
+
 function show_password($gid, $level) {
 	$results = get_my_passwords($gid);
 	foreach ($results as $entry) {
@@ -24,6 +25,8 @@ function show_password($gid, $level) {
 		$name = $entry['name'];
 		$login = decrypt_string($entry['login']);
 		$notes = $entry['note'];
+		$myRights = user_rights();
+		
 		?>
 							<tr onclick="window.location='show.php?id=<?php echo $id;?>'">
 								<td><?php echo str_repeat('&nbsp;', $level *4 +4).$name; ?></td>
@@ -31,8 +34,12 @@ function show_password($gid, $level) {
 								<td><?php echo $notes; ?></td>
 								<td>
 									<a class="btn btn-primary btn-xs" href="show.php?id=<?php echo $id;?>">Show</a>
-									<a class="btn btn-primary btn-xs" href="edit.php?id=<?php echo $id;?>">Edit</a>
-									<a class="btn btn-danger btn-xs" href="delete.php?id=<?php echo $id;?>">Delete</a>
+									<?php if (($myRights & accPasswordEdit) != 0) { ?>
+										<a class="btn btn-primary btn-xs" href="edit.php?id=<?php echo $id;?>">Edit</a>
+									<?php } ?>
+									<?php if (($myRights & accPasswordRemove) != 0) { ?>
+										<a class="btn btn-danger btn-xs" href="delete.php?id=<?php echo $id;?>">Delete</a>
+									<?php } ?>
 								</td>
 							</tr>
 		<?php
@@ -42,15 +49,17 @@ function show_password($gid, $level) {
 function show_group($gid, $level = 0) {
 	$groups = get_group_list($gid);
 	foreach ($groups as $group) {
-		$id = $group['gid'];
-		$path = get_group_path($group['gid']);
-		$description = get_group_description($group['gid']);
-		
-		echo "<tr style=\"background-color:#d6eaf8; padding:0\"><td style=\"padding-top: 0;padding-bottom: 0\"><h5 style \" margin-top:0\"><b>".str_repeat('&nbsp;', $level *4).$path."</b></h5></td><td></td><td>".$description."</td><td></td>";
-		
-		show_password($id, $level);
-		
-		show_group($id, $level +1);
+		if (check_group_permissions($group['gid'], get_my_uid()))  {
+			$id = $group['gid'];
+			$path = get_group_path($group['gid']);
+			$description = get_group_description($group['gid']);
+			
+			echo "<tr style=\"background-color:#d6eaf8; padding:0\"><td style=\"padding-top: 0;padding-bottom: 0\"><h5 style \" margin-top:0\"><b>".str_repeat('&nbsp;', $level *4).$path."</b></h5></td><td></td><td>".$description."</td><td></td>";
+			
+			show_password($id, $level);
+			
+			show_group($id, $level +1);
+		}
 	}
 }
